extract into platform-specific modules for a bit of maintainability
